+++ framework-crates/objc2-metal/src/packed.rs
@@ -1,42 +1,18 @@
-use core::fmt;
 use std::os::raw::c_float;
 
 use objc2::encode::{Encode, Encoding, RefEncode};
 
-// TODO: Investigate the need for this helper struct?
-//
-// I'm fairly sure that just replacing `MTLPackedFloat3` with this struct has
-// the same effect in Rust as doing the trick with putting it in a union?
-//
-// Or is the intention actually to create something similar to `#[packed]`?
+// `MTLPackedFloat3` is actually a union internally, but replacing it with
+// this simpler struct same semantics in Rust, and is generally the nicer
+// pattern.
 //
 // <https://users.rust-lang.org/t/mapping-nested-packed-union-from-c-to-rust/87334>
-
 #[repr(C)]
 #[derive(Clone, Copy, Debug, PartialEq)]
-pub struct __MTLPackedFloat3 {
-    x: c_float,
-    y: c_float,
-    z: c_float,
-}
-
-#[repr(C)]
-#[derive(Clone, Copy)]
-pub union MTLPackedFloat3 {
-    struct_: __MTLPackedFloat3,
-    elements: [c_float; 3],
-}
-
-impl PartialEq for MTLPackedFloat3 {
-    fn eq(&self, other: &Self) -> bool {
-        unsafe { self.struct_ == other.struct_ }
-    }
-}
-
-impl fmt::Debug for MTLPackedFloat3 {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        unsafe { self.struct_.fmt(f) }
-    }
+pub struct MTLPackedFloat3 {
+    pub x: c_float,
+    pub y: c_float,
+    pub z: c_float,
 }
 
 unsafe impl Encode for MTLPackedFloat3 {